requirements.txt
weaviate-client
python_dotenv
fastapi
uvicorn
git+https://github.com/weaviate/weaviate-recommend-python-client.git
pandas